Port to Baystation 12

+++ code/game/gamemodes/meteor/meteors.dm
@@ -1,45 +1,89 @@
-#define DEFAULT_METEOR_LIFETIME 1800
-GLOBAL_VAR_INIT(meteor_wave_delay, 625) //minimum wait between waves in tenths of seconds
+/var/const/meteor_wave_delay = 1 MINUTE //minimum wait between waves in tenths of seconds
 //set to at least 100 unless you want evarr ruining every round
 
-//Meteors probability of spawning during a given wave
-GLOBAL_LIST_INIT(meteors_normal, list(/obj/effect/meteor/dust=3, /obj/effect/meteor/medium=8, /obj/effect/meteor/big=3, \
-						  /obj/effect/meteor/flaming=1, /obj/effect/meteor/irradiated=3)) //for normal meteor event
+//Meteor groups, used for various random events and the Meteor gamemode.
 
-GLOBAL_LIST_INIT(meteors_threatening, list(/obj/effect/meteor/medium=4, /obj/effect/meteor/big=8, \
-						  /obj/effect/meteor/flaming=3, /obj/effect/meteor/irradiated=3)) //for threatening meteor event
+// Dust, used by space dust event and during earliest stages of meteor mode.
+/var/list/meteors_dust = list(/obj/effect/meteor/dust)
 
-GLOBAL_LIST_INIT(meteors_catastrophic, list(/obj/effect/meteor/medium=5, /obj/effect/meteor/big=75, \
-						  /obj/effect/meteor/flaming=10, /obj/effect/meteor/irradiated=10, /obj/effect/meteor/tunguska = 1)) //for catastrophic meteor event
+// Standard meteors, used during early stages of the meteor gamemode.
+/var/list/meteors_normal = list(\
+		/obj/effect/meteor/medium=8,\
+		/obj/effect/meteor/dust=3,\
+		/obj/effect/meteor/irradiated=3,\
+		/obj/effect/meteor/big=3,\
+		/obj/effect/meteor/flaming=1,\
+		/obj/effect/meteor/golden=1,\
+		/obj/effect/meteor/silver=1\
+		)
 
-GLOBAL_LIST_INIT(meteorsB, list(/obj/effect/meteor/meaty=5, /obj/effect/meteor/meaty/xeno=1)) //for meaty ore event
+// Threatening meteors, used during the meteor gamemode.
+/var/list/meteors_threatening = list(\
+		/obj/effect/meteor/big=10,\
+		/obj/effect/meteor/medium=5,\
+		/obj/effect/meteor/golden=3,\
+		/obj/effect/meteor/silver=3,\
+		/obj/effect/meteor/flaming=3,\
+		/obj/effect/meteor/irradiated=3,\
+		/obj/effect/meteor/emp=3\
+		)
 
-GLOBAL_LIST_INIT(meteorsC, list(/obj/effect/meteor/dust)) //for space dust event
+// Catastrophic meteors, pretty dangerous without shields and used during the meteor gamemode.
+/var/list/meteors_catastrophic = list(\
+		/obj/effect/meteor/big=75,\
+		/obj/effect/meteor/flaming=10,\
+		/obj/effect/meteor/irradiated=10,\
+		/obj/effect/meteor/emp=10,\
+		/obj/effect/meteor/medium=5,\
+		/obj/effect/meteor/golden=4,\
+		/obj/effect/meteor/silver=4,\
+		/obj/effect/meteor/tunguska=1\
+		)
+
+// Armageddon meteors, very dangerous, and currently used only during the meteor gamemode.
+/var/list/meteors_armageddon = list(\
+		/obj/effect/meteor/big=25,\
+		/obj/effect/meteor/flaming=10,\
+		/obj/effect/meteor/irradiated=10,\
+		/obj/effect/meteor/emp=10,\
+		/obj/effect/meteor/medium=3,\
+		/obj/effect/meteor/tunguska=3,\
+		/obj/effect/meteor/golden=2,\
+		/obj/effect/meteor/silver=2\
+		)
+
+// Cataclysm meteor selection. Very very dangerous and effective even against shields. Used in late game meteor gamemode only.
+/var/list/meteors_cataclysm = list(\
+		/obj/effect/meteor/big=40,\
+		/obj/effect/meteor/emp=20,\
+		/obj/effect/meteor/tunguska=20,\
+		/obj/effect/meteor/irradiated=10,\
+		/obj/effect/meteor/golden=10,\
+		/obj/effect/meteor/silver=10,\
+		/obj/effect/meteor/flaming=10,\
+		/obj/effect/meteor/supermatter=1\
+		)
+
 
 
 ///////////////////////////////
 //Meteor spawning global procs
 ///////////////////////////////
 
-/proc/spawn_meteors(number = 10, list/meteortypes)
+/proc/spawn_meteors(var/number = 10, var/list/meteortypes, var/startSide, var/zlevel)
 	for(var/i = 0; i < number; i++)
-		spawn_meteor(meteortypes)
+		spawn_meteor(meteortypes, startSide, zlevel)
 
-/proc/spawn_meteor(list/meteortypes)
-	var/turf/pickedstart
-	var/turf/pickedgoal
-	var/max_i = 10//number of tries to spawn meteor.
-	while(!isspaceturf(pickedstart))
-		var/startSide = pick(GLOB.cardinals)
-		var/startZ = pick(SSmapping.levels_by_trait(ZTRAIT_STATION))
-		pickedstart = spaceDebrisStartLoc(startSide, startZ)
-		pickedgoal = spaceDebrisFinishLoc(startSide, startZ)
-		max_i--
-		if(max_i<=0)
-			return
+/proc/spawn_meteor(var/list/meteortypes, var/startSide, var/zlevel)
+	var/turf/pickedstart = spaceDebrisStartLoc(startSide, zlevel)
+	var/turf/pickedgoal = spaceDebrisFinishLoc(startSide, zlevel)
+
 	var/Me = pickweight(meteortypes)
-	var/obj/effect/meteor/M = new Me(pickedstart, pickedgoal)
+	var/obj/effect/meteor/M = new Me(pickedstart)
 	M.dest = pickedgoal
+	spawn(0)
+		walk_towards(M, M.dest, 1)
+	return
 
 /proc/spaceDebrisStartLoc(startSide, Z)
 	var/starty
@@ -57,25 +101,27 @@ GLOBAL_LIST_INIT(meteorsC, list(/obj/effect/meteor/dust)) //for space dust event
 		if(WEST)
 			starty = rand((TRANSITIONEDGE+1), world.maxy-(TRANSITIONEDGE+1))
 			startx = (TRANSITIONEDGE+1)
-	. = locate(startx, starty, Z)
+	var/turf/T = locate(startx, starty, Z)
+	return T
 
 /proc/spaceDebrisFinishLoc(startSide, Z)
 	var/endy
 	var/endx
 	switch(startSide)
 		if(NORTH)
-			endy = (TRANSITIONEDGE+1)
-			endx = rand((TRANSITIONEDGE+1), world.maxx-(TRANSITIONEDGE+1))
+			endy = TRANSITIONEDGE
+			endx = rand(TRANSITIONEDGE, world.maxx-TRANSITIONEDGE)
 		if(EAST)
-			endy = rand((TRANSITIONEDGE+1), world.maxy-(TRANSITIONEDGE+1))
-			endx = (TRANSITIONEDGE+1)
+			endy = rand(TRANSITIONEDGE, world.maxy-TRANSITIONEDGE)
+			endx = TRANSITIONEDGE
 		if(SOUTH)
-			endy = world.maxy-(TRANSITIONEDGE+1)
-			endx = rand((TRANSITIONEDGE+1), world.maxx-(TRANSITIONEDGE+1))
+			endy = world.maxy-TRANSITIONEDGE
+			endx = rand(TRANSITIONEDGE, world.maxx-TRANSITIONEDGE)
 		if(WEST)
-			endy = rand((TRANSITIONEDGE+1),world.maxy-(TRANSITIONEDGE+1))
-			endx = world.maxx-(TRANSITIONEDGE+1)
-	. = locate(endx, endy, Z)
+			endy = rand(TRANSITIONEDGE,world.maxy-TRANSITIONEDGE)
+			endx = world.maxx-TRANSITIONEDGE
+	var/turf/T = locate(endx, endy, Z)
+	return T
 
 ///////////////////////
 //The meteor effect
@@ -86,71 +132,64 @@ GLOBAL_LIST_INIT(meteorsC, list(/obj/effect/meteor/dust)) //for space dust event
 	desc = "You should probably run instead of gawking at this."
 	icon = 'icons/obj/meteor.dmi'
 	icon_state = "small"
-	density = TRUE
-	anchored = TRUE
+	density = 1
+	anchored = 1
 	var/hits = 4
 	var/hitpwr = 2 //Level of ex_act to be called on hit.
 	var/dest
-	pass_flags = PASSTABLE
+	pass_flags = PASS_FLAG_TABLE
 	var/heavy = 0
-	var/meteorsound = 'sound/effects/meteorimpact.ogg'
 	var/z_original
-	var/threat = 0 // used for determining which meteors are most interesting
-	var/lifetime = DEFAULT_METEOR_LIFETIME
-	var/timerid = null
-	var/list/meteordrop = list(/obj/item/stack/ore/iron)
-	var/dropamt = 2
+	var/meteordrop = /obj/item/weapon/ore/iron
+	var/dropamt = 1
+	var/ismissile //missiles don't spin
+
+	var/move_count = 0
+
+/obj/effect/meteor/proc/get_shield_damage()
+	return max(((max(hits, 2)) * (heavy + 1) * rand(30, 60)) / hitpwr , 0)
+
+/obj/effect/meteor/New()
+	..()
+	z_original = z
 
 /obj/effect/meteor/Move()
-	if(z != z_original || loc == dest)
-		qdel(src)
-		return FALSE
-
 	. = ..() //process movement...
+	move_count++
+	if(loc == dest)
+		qdel(src)
 
-	if(.)//.. if did move, ram the turf we get in
-		var/turf/T = get_turf(loc)
-		ram_turf(T)
-
-		if(prob(10) && !isspaceturf(T))//randomly takes a 'hit' from ramming
-			get_hit()
+/obj/effect/meteor/touch_map_edge()
+	if(move_count > TRANSITIONEDGE)
+		qdel(src)
 
 /obj/effect/meteor/Destroy()
-	if (timerid)
-		deltimer(timerid)
-	GLOB.meteor_list -= src
-	SSaugury.unregister_doom(src)
 	walk(src,0) //this cancels the walk_towards() proc
-	. = ..()
+	return ..()
 
-/obj/effect/meteor/Initialize(mapload, target)
-	. = ..()
-	z_original = z
-	GLOB.meteor_list += src
-	SSaugury.register_doom(src, threat)
-	SpinAnimation()
-	timerid = QDEL_IN(src, lifetime)
-	chase_target(target)
+/obj/effect/meteor/New()
+	..()
+	if(!ismissile)
+		SpinAnimation()
 
-/obj/effect/meteor/Collide(atom/A)
-	if(A)
+/obj/effect/meteor/Bump(atom/A)
+	..()
+	if(A && !QDELETED(src))	// Prevents explosions and other effects when we were deleted by whatever we Bumped() - currently used by shields.
 		ram_turf(get_turf(A))
-		playsound(src.loc, meteorsound, 40, 1)
-		get_hit()
+		get_hit() //should only get hit once per move attempt
 
-/obj/effect/meteor/proc/ram_turf(turf/T)
+/obj/effect/meteor/CanPass(atom/movable/mover, turf/target, height=0, air_group=0)
+	return istype(mover, /obj/effect/meteor) ? 1 : ..()
+
+/obj/effect/meteor/proc/ram_turf(var/turf/T)
 	//first bust whatever is in the turf
 	for(var/atom/A in T)
-		if(A != src)
-			if(isliving(A))
-				A.visible_message("<span class='warning'>[src] slams into [A].</span>", "<span class='userdanger'>[src] slams into you!.</span>")
+		if(A != src && !A.CanPass(src, src.loc, 0.5, 0)) //only ram stuff that would actually block us
 			A.ex_act(hitpwr)
 
 	//then, ram the turf if it still exists
-	if(T)
+	if(T && !T.CanPass(src, src.loc, 0.5, 0))
 		T.ex_act(hitpwr)
-
-
 
 //process getting 'hit' by colliding with a dense object
 //or randomly when ramming turfs
@@ -164,42 +203,26 @@ GLOBAL_LIST_INIT(meteorsC, list(/obj/effect/meteor/dust)) //for space dust event
 /obj/effect/meteor/ex_act()
 	return
 
-/obj/effect/meteor/examine(mob/user)
-	if(!(flags_1 & ADMIN_SPAWNED_1) && isliving(user))
-		SSmedals.UnlockMedal(MEDAL_METEOR, user.client)
-	..()
-
-/obj/effect/meteor/attackby(obj/item/I, mob/user, params)
-	if(I.tool_behaviour == TOOL_MINING)
-		make_debris()
+/obj/effect/meteor/attackby(obj/item/weapon/W as obj, mob/user as mob, params)
+	if(istype(W, /obj/item/weapon/pickaxe))
 		qdel(src)
-	else
-		. = ..()
+		return
+	..()
 
 /obj/effect/meteor/proc/make_debris()
 	for(var/throws = dropamt, throws > 0, throws--)
-		var/thing_to_spawn = pick(meteordrop)
-		new thing_to_spawn(get_turf(src))
-
-/obj/effect/meteor/proc/chase_target(atom/chasing, delay = 1)
-	set waitfor = FALSE
-	if(chasing)
-		walk_towards(src, chasing, delay)
+		var/obj/item/O = new meteordrop(get_turf(src))
+		O.throw_at(dest, 5, 10)
 
 /obj/effect/meteor/proc/meteor_effect()
 	if(heavy)
-		var/sound/meteor_sound = sound(meteorsound)
-		var/random_frequency = get_rand_frequency()
-
 		for(var/mob/M in GLOB.player_list)
-			if((M.orbiting) && (SSaugury.watchers[M]))
-				continue
 			var/turf/T = get_turf(M)
 			if(!T || T.z != src.z)
 				continue
 			var/dist = get_dist(M.loc, src.loc)
-			shake_camera(M, dist > 20 ? 2 : 4, dist > 20 ? 1 : 3)
-			M.playsound_local(src.loc, null, 50, 1, random_frequency, 10, S = meteor_sound)
+			shake_camera(M, dist > 20 ? 3 : 5, dist > 20 ? 1 : 3)
+
 
 ///////////////////////
 //Meteor types
@@ -209,18 +232,16 @@ GLOBAL_LIST_INIT(meteorsC, list(/obj/effect/meteor/dust)) //for space dust event
 /obj/effect/meteor/dust
 	name = "space dust"
 	icon_state = "dust"
-	pass_flags = PASSTABLE | PASSGRILLE
+	pass_flags = PASS_FLAG_TABLE | PASS_FLAG_GRILLE
 	hits = 1
 	hitpwr = 3
-	meteorsound = 'sound/weapons/gunshot_smg.ogg'
-	meteordrop = list(/obj/item/stack/ore/glass)
-	threat = 1
+	dropamt = 1
+	meteordrop = /obj/item/weapon/ore/glass
 
 //Medium-sized
 /obj/effect/meteor/medium
 	name = "meteor"
-	dropamt = 3
-	threat = 5
+	dropamt = 2
 
 /obj/effect/meteor/medium/meteor_effect()
 	..()
@@ -228,12 +249,11 @@ GLOBAL_LIST_INIT(meteorsC, list(/obj/effect/meteor/dust)) //for space dust event
 
 //Large-sized
 /obj/effect/meteor/big
-	name = "big meteor"
+	name = "large meteor"
 	icon_state = "large"
 	hits = 6
 	heavy = 1
-	dropamt = 4
-	threat = 10
+	dropamt = 3
 
 /obj/effect/meteor/big/meteor_effect()
 	..()
@@ -245,9 +265,7 @@ GLOBAL_LIST_INIT(meteorsC, list(/obj/effect/meteor/dust)) //for space dust event
 	icon_state = "flaming"
 	hits = 5
 	heavy = 1
-	meteorsound = 'sound/effects/bamf.ogg'
-	meteordrop = list(/obj/item/stack/ore/plasma)
-	threat = 20
+	meteordrop = /obj/item/weapon/ore/phoron
 
 /obj/effect/meteor/flaming/meteor_effect()
 	..()
@@ -258,112 +276,114 @@ GLOBAL_LIST_INIT(meteorsC, list(/obj/effect/meteor/dust)) //for space dust event
 	name = "glowing meteor"
 	icon_state = "glowing"
 	heavy = 1
-	meteordrop = list(/obj/item/stack/ore/uranium)
-	threat = 15
-
+	meteordrop = /obj/item/weapon/ore/uranium
 
 /obj/effect/meteor/irradiated/meteor_effect()
 	..()
 	explosion(src.loc, 0, 0, 4, 3, 0)
 	new /obj/effect/decal/cleanable/greenglow(get_turf(src))
-	radiation_pulse(src, 500)
+	SSradiation.radiate(src, 50)
 
-//Meaty Ore
-/obj/effect/meteor/meaty
-	name = "meaty ore"
-	icon_state = "meateor"
-	desc = "Just... don't think too hard about where this thing came from."
-	hits = 2
-	heavy = 1
-	meteorsound = 'sound/effects/blobattack.ogg'
-	meteordrop = list(/obj/item/reagent_containers/food/snacks/meat/slab/human, /obj/item/reagent_containers/food/snacks/meat/slab/human/mutant, /obj/item/organ/heart, /obj/item/organ/lungs, /obj/item/organ/tongue, /obj/item/organ/appendix/)
-	var/meteorgibs = /obj/effect/gibspawner/generic
-	threat = 2
+/obj/effect/meteor/golden
+	name = "golden meteor"
+	icon_state = "glowing"
+	desc = "Shiny! But also deadly."
+	meteordrop = /obj/item/weapon/ore/gold
 
-/obj/effect/meteor/meaty/Initialize()
-	for(var/path in meteordrop)
-		if(path == /obj/item/reagent_containers/food/snacks/meat/slab/human/mutant)
-			meteordrop -= path
-			meteordrop += pick(subtypesof(path))
+/obj/effect/meteor/silver
+	name = "silver meteor"
+	icon_state = "glowing_blue"
+	desc = "Shiny! But also deadly."
+	meteordrop = /obj/item/weapon/ore/silver
 
-	for(var/path in meteordrop)
-		if(path == /obj/item/organ/tongue)
-			meteordrop -= path
-			meteordrop += pick(typesof(path))
-	return ..()
+/obj/effect/meteor/emp
+	name = "conducting meteor"
+	icon_state = "glowing_blue"
+	desc = "Hide your floppies!"
+	meteordrop = /obj/item/weapon/ore/osmium
+	dropamt = 2
 
-/obj/effect/meteor/meaty/make_debris()
+/obj/effect/meteor/emp/meteor_effect()
 	..()
-	new meteorgibs(get_turf(src))
+	// Best case scenario: Comparable to a low-yield EMP grenade.
+	// Worst case scenario: Comparable to a standard yield EMP grenade.
+	empulse(src, rand(2, 4), rand(4, 10))
 
-
-/obj/effect/meteor/meaty/ram_turf(turf/T)
-	if(!isspaceturf(T))
-		new /obj/effect/decal/cleanable/blood(T)
-
-/obj/effect/meteor/meaty/Collide(atom/A)
-	A.ex_act(hitpwr)
-	get_hit()
-
-//Meaty Ore Xeno edition
-/obj/effect/meteor/meaty/xeno
-	color = "#5EFF00"
-	meteordrop = list(/obj/item/reagent_containers/food/snacks/meat/slab/xeno, /obj/item/organ/tongue/alien)
-	meteorgibs = /obj/effect/gibspawner/xeno
-
-/obj/effect/meteor/meaty/xeno/Initialize()
-	meteordrop += subtypesof(/obj/item/organ/alien)
-	return ..()
-
-/obj/effect/meteor/meaty/xeno/ram_turf(turf/T)
-	if(!isspaceturf(T))
-		new /obj/effect/decal/cleanable/xenoblood(T)
+/obj/effect/meteor/emp/get_shield_damage()
+	return ..() * rand(2,4)
 
 //Station buster Tunguska
 /obj/effect/meteor/tunguska
 	name = "tunguska meteor"
 	icon_state = "flaming"
 	desc = "Your life briefly passes before your eyes the moment you lay them on this monstrosity."
-	hits = 30
+	hits = 10
 	hitpwr = 1
 	heavy = 1
-	meteorsound = 'sound/effects/bamf.ogg'
-	meteordrop = list(/obj/item/stack/ore/plasma)
-	threat = 50
-
-/obj/effect/meteor/tunguska/Move()
-	. = ..()
-	if(.)
-		new /obj/effect/temp_visual/revenant(get_turf(src))
+	meteordrop = /obj/item/weapon/ore/diamond	// Probably means why it penetrates the hull so easily before exploding.
 
 /obj/effect/meteor/tunguska/meteor_effect()
 	..()
-	explosion(src.loc, 5, 10, 15, 20, 0)
+	explosion(src.loc, 3, 6, 9, 20, 0)
 
-/obj/effect/meteor/tunguska/Collide()
+// This is the final solution against shields - a single impact can bring down most shield generators.
+/obj/effect/meteor/supermatter
+	name = "supermatter shard"
+	desc = "Oh god, what will be next..?"
+	icon = 'icons/obj/engine.dmi'
+	icon_state = "darkmatter"
+
+/obj/effect/meteor/supermatter/meteor_effect()
 	..()
-	if(prob(20))
-		explosion(src.loc,2,4,6,8)
+	explosion(src.loc, 1, 2, 3, 4, 0)
+	for(var/obj/machinery/power/apc/A in range(rand(12, 20), src))
+		A.energy_fail(round(10 * rand(8, 12)))
 
-//////////////////////////
-//Spookoween meteors
-/////////////////////////
+/obj/effect/meteor/supermatter/get_shield_damage()
+	return ..() * rand(80, 120)
 
-GLOBAL_LIST_INIT(meteorsSPOOKY, list(/obj/effect/meteor/pumpkin))
+//Missiles, for events and so on
+/obj/effect/meteor/supermatter/missile
+	name = "photon torpedo"
+	desc = "An advanded warhead designed to tactically destroy space installations."
+	icon = 'icons/obj/missile.dmi'
+	icon_state = "photon"
+	meteordrop = null
+	ismissile = TRUE
+	dropamt = 0
 
-/obj/effect/meteor/pumpkin
-	name = "PUMPKING"
-	desc = "THE PUMPKING'S COMING!"
-	icon = 'icons/obj/meteor_spooky.dmi'
-	icon_state = "pumpkin"
-	hits = 10
-	heavy = 1
-	dropamt = 1
-	meteordrop = list(/obj/item/clothing/head/hardhat/pumpkinhead, /obj/item/reagent_containers/food/snacks/grown/pumpkin)
-	threat = 100
+/obj/effect/meteor/medium/missile
+	name = "missile"
+	desc = "Some kind of missile."
+	icon = 'icons/obj/missile.dmi'
+	icon_state = "missile"
+	meteordrop = null
+	ismissile = TRUE
+	dropamt = 0
 
-/obj/effect/meteor/pumpkin/Initialize()
-	. = ..()
-	meteorsound = pick('sound/hallucinations/im_here1.ogg','sound/hallucinations/im_here2.ogg')
-//////////////////////////
-#undef DEFAULT_METEOR_LIFETIME
+/obj/effect/meteor/big/missile
+	name = "high-yield missile"
+	desc = "Some kind of missile."
+	icon = 'icons/obj/missile.dmi'
+	icon_state = "missile"
+	meteordrop = null
+	ismissile = TRUE
+	dropamt = 0
+
+/obj/effect/meteor/flaming/missile
+	name = "incendiary missile"
+	desc = "Some kind of missile."
+	icon = 'icons/obj/missile.dmi'
+	icon_state = "missile"
+	meteordrop = null
+	ismissile = TRUE
+	dropamt = 0
+
+/obj/effect/meteor/emp/missile
+	name = "ion torpedo"
+	desc = "Some kind of missile."
+	icon = 'icons/obj/missile.dmi'
+	icon_state = "torpedo"
+	meteordrop = null
+	ismissile = TRUE
+	dropamt = 0
